requirements.txt
## The following requirements were added by pip freeze:
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.34.0
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
newsapi-python==0.2.7
Pillow==10.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
whitenoise==6.5.0
gunicorn
psycopg2
